moved module f_set_metadat to mgk_file_handling.py . Changes to uploader
